reoranize commands
