+++ wvlet-cli/src/main/scala/wvlet/lang/cli/WvletGlobalOption.scala
@@ -1,0 +1,36 @@
+package wvlet.lang.cli
+
+import wvlet.airframe.launcher.option
+import wvlet.lang.BuildInfo
+import wvlet.log.{LogLevel, LogSupport, Logger}
+
+case class WvletGlobalOption(
+    @option(prefix = "-h,--help", description = "Display help message", isHelp = true)
+    displayHelp: Boolean = false,
+    @option(prefix = "--debug", description = "Enable debug log")
+    debugMode: Boolean = false,
+    @option(prefix = "-l", description = "log level")
+    logLevel: LogLevel = LogLevel.INFO,
+    @option(prefix = "-L", description = "log level for a class pattern")
+    logLevelPatterns: List[String] = List.empty
+) extends LogSupport:
+  Logger("wvlet.lang.runner").setLogLevel {
+    if debugMode then
+      LogLevel.DEBUG
+    else
+      logLevel
+  }
+
+  def versionString = s"wvlet version: ${BuildInfo.version} (Built at: ${BuildInfo.builtAtString})"
+  debug(versionString)
+
+  logLevelPatterns.foreach { p =>
+    p.split("=") match
+      case Array(pattern, level) =>
+        debug(s"Set the log level for ${pattern} to ${level}")
+        Logger.setLogLevel(pattern, LogLevel(level))
+      case _ =>
+        error(s"Invalid log level pattern: ${p}")
+  }
+
+end WvletGlobalOption
