+++ wvlet-cli/src/test/scala/wvlet/lang/cli/WvMainTest.scala
@@ -15,46 +15,45 @@ package wvlet.lang.cli
 
 import wvlet.airspec.AirSpec
 
-class WvletREPLTest extends AirSpec:
+class WvMainTest extends AirSpec:
   test("help") {
-    WvletCli.main("-c 'help'")
+    WvMain.main("-c 'help'")
   }
 
   test("model in the working folder") {
-    WvletCli.main("-w spec/model1 -c 'from person_filter(2)'")
+    WvMain.main("-w spec/model1 -c 'from person_filter(2)'")
   }
 
   test("def new model") {
-    WvletCli
-      .main("-w spec/model1 -c 'model m(v:int) = from person where id = v end' -c 'from m(1)'")
+    WvMain.main("-w spec/model1 -c 'model m(v:int) = from person where id = v end' -c 'from m(1)'")
   }
 
   test("show models") {
-    WvletCli.main("-w spec/model1 -c 'show models'")
-    WvletCli.main("-w spec/model1 -c 'show models limit 1'")
-    WvletCli.main("-w spec/model1 -c 'show models' -c 'show models limit 5'")
+    WvMain.main("-w spec/model1 -c 'show models'")
+    WvMain.main("-w spec/model1 -c 'show models limit 1'")
+    WvMain.main("-w spec/model1 -c 'show models' -c 'show models limit 5'")
   }
 
   test("select group index") {
-    WvletCli.main("-w spec/model1 -c 'from person group by age / 10 select _1'")
+    WvMain.main("-w spec/model1 -c 'from person group by age / 10 select _1'")
   }
 
   test("clip") {
-    WvletCli.main("-w spec/model1 -c 'from person' -c 'clip'")
+    WvMain.main("-w spec/model1 -c 'from person' -c 'clip'")
   }
 
   test("clip-result") {
-    WvletCli.main("-w spec/model1 -c 'from person' -c 'clip-result'")
+    WvMain.main("-w spec/model1 -c 'from person' -c 'clip-result'")
   }
 
   test("rows") {
-    WvletCli.main("-w spec/model1 -c 'rows 2' -c 'from person'")
+    WvMain.main("-w spec/model1 -c 'rows 2' -c 'from person'")
   }
 
   test("col-width") {
-    WvletCli.main(
+    WvMain.main(
       """-c 'col-width 10' -c "from 'https://shell.duckdb.org/data/tpch/0_01/parquet/customer.parquet'""""
     )
   }
 
-end WvletREPLTest
+end WvMainTest
